# Improve lexicographic comparison of small integers