This PR adds an improved implementation of lexicographic comparison of 32-bit integers as natively implemented in V8:

https://github.com/v8/v8/blob/8e02f47/src/runtime/runtime-numbers.cc#L197-L270

Unlike the V8 implementation, the implementation of `log10` estimation relies solely on comparisons whereas V8 uses a bit twiddling hack. My guess is that this is why the pure-JS implementation is in fact able to outperform V8.