@kasperisager 2016-06-11
Just ran the benchmark using the lexicographic comparators and the numbers are looking pretty good.

## Before

```console
ArrayType                     Length TimSort   array.sort  Speedup
randomInt                     10     1361      2153        1.5821247181403686
randomInt                     100    28718     28846       1.004459231998948
randomInt                     1000   499297    506045      1.0135140276678558
randomInt                     10000  10526157  13812425    1.3122001444774758
descendingInt                 10     8535      15053       1.7636446026090264
descendingInt                 100    110867    268601      2.4227276735432066
descendingInt                 1000   616190    2887679     4.686340037840632
descendingInt                 10000  5145321   32137367    6.2459397041643765
ascendingInt                  10     2805      2284        0.8144553782965919
ascendingInt                  100    85724     222489      2.59539265566043
ascendingInt                  1000   595666    2624008     4.405160310773195
ascendingInt                  10000  4628419   29136848    6.295204524096742
ascending3RandomExchangesInt  10     5968      4448        0.7452532923961506
ascending3RandomExchangesInt  100    103889    224566      2.161578082970083
ascending3RandomExchangesInt  1000   565668    2701604     4.775945986406893
ascending3RandomExchangesInt  10000  4530521   28687487    6.332050259434433
ascending10RandomEndInt       10     5775      7942        1.3752085339868254
ascending10RandomEndInt       100    116257    238025      2.0474035217754447
ascending10RandomEndInt       1000   576414    2674571     4.640014226952042
ascending10RandomEndInt       10000  4667820   28427639    6.090131310284939
allEqualInt                   10     163       789         4.83007987830857
allEqualInt                   100    550       2270        4.121793762113859
allEqualInt                   1000   2867      18579       6.478362281312396
allEqualInt                   10000  27065     206801      7.640737275451406
manyDuplicateInt              10     4575      6490        1.418540635041049
manyDuplicateInt              100    199369    215305      1.0799307757354586
manyDuplicateInt              1000   2100510   2364865     1.1258525682645177
manyDuplicateInt              10000  21107902  24394843    1.155720844307284
someDuplicateInt              10     5820      7960        1.3675482063217754
someDuplicateInt              100    199566    215040      1.0775359732318404
someDuplicateInt              1000   2210138   2455562     1.111044558298126
someDuplicateInt              10000  21932111  24341676    1.1098646951174247
```

## After

```console
ArrayType                     Length TimSort   array.sort  Speedup
randomInt                     10     1485      2150        1.4481724572093047
randomInt                     100    31922     28722       0.8997696527001674
randomInt                     1000   555997    506724      0.9113788708295715
randomInt                     10000  11311205  14261782    1.2608543418624758
descendingInt                 10     1037      15153       14.60138031843942
descendingInt                 100    10880     260443      23.937659538803995
descendingInt                 1000   102430    2825886     27.588380991868892
descendingInt                 10000  1148489   30964060    26.960679817628066
ascendingInt                  10     417       2308        5.524879354257732
ascendingInt                  100    9840      218797      22.233793914116408
ascendingInt                  1000   81024     2558330     31.57495135110461
ascendingInt                  10000  926406    27766735    29.972508991397596
ascending3RandomExchangesInt  10     732       4388        5.989348243010266
ascending3RandomExchangesInt  100    12203     223560      18.319676548953012
ascending3RandomExchangesInt  1000   80104     2735095     34.14410272647065
ascending3RandomExchangesInt  10000  934359    27258835    29.173812193823977
ascending10RandomEndInt       10     904       8008        8.856150662969483
ascending10RandomEndInt       100    13605     235026      17.274426619353832
ascending10RandomEndInt       1000   83932     2642180     31.479905629390448
ascending10RandomEndInt       10000  826108    27000964    32.684528927659336
allEqualInt                   10     227       749         3.2920186560009457
allEqualInt                   100    1176      2572        2.1866399576772393
allEqualInt                   1000   9526      19368       2.0331083289329483
allEqualInt                   10000  106558    184289      1.7294583379030581
manyDuplicateInt              10     843       6517        7.7255778236620785
manyDuplicateInt              100    23121     211967      9.167336691650306
manyDuplicateInt              1000   373416    2263669     6.062048491246288
manyDuplicateInt              10000  4937504   25134781    5.0905835259694285
someDuplicateInt              10     886       8104        9.142387068611173
someDuplicateInt              100    23109     214847      9.296760122090921
someDuplicateInt              1000   364971    2362090     6.4719881857106705
someDuplicateInt              10000  5100654   24832775    4.868547394014137
```