@mziccard 2016-06-11
@kasperisager This seems like super good news. I'll try to go through your changes as soon as possible. Just to be sure, which version of node did you run the benchmarks against?