@kasperisager 2016-06-12
Sounds good, cheers! Ran the benchmark on `v6.1.0`.