@kasperisager 2016-07-15
Have you had a chance to look this over yet?