@mziccard 2016-07-21
@kasperisager sorry it took me ages to go through this. Everything looks good on my side. We should probably remove the following performance note from the README:

> array.sort will probably still be faster at lexicographically sorting arrays of numbers. In this case, the timsort module inefficiently converts values to strings inside the compare function and then compares the strings. array.sort, instead, uses a smarter and faster lexicographic comparison of numbers (will try to do something similar soon).

I can take care of doing the above README change. I should also update the benchmarks to a more recent version of node.