@mziccard 2016-07-24
@kasperisager I finally merged your change and released a new version (0.3.0). Thanks again!